int DBConfig

+++ BL/LaptopBL.cs
@@ -1,4 +1,3 @@
-using System;
 using Persistance;
 using DAL;
 using System.Collections.Generic;
@@ -8,7 +7,7 @@ namespace BL
 {
     public class LaptopBL
     {
-        
+
         private LaptopDAL dal = new LaptopDAL();
         private LaptopsDAL sdal = new LaptopsDAL();
 
@@ -47,59 +46,35 @@ namespace BL
             {
                 return true;
             }
-            // var hasNumber = new Regex(@"[0-9]+");
-            // var hasUpperChar = new Regex(@"[A-Z]+");
-            // var hasMiniMaxChars = new Regex(@".{8,15}");
-            // var hasLowerChar = new Regex(@"[a-z]+");
-
-            // if (!hasUpperChar.IsMatch(input) && !hasLowerChar.IsMatch(input))
-            // {
-            // ErrorMessage = "Name should contain a letter";
-            // return false;
-            // }
-            // else if (!hasMiniMaxChars.IsMatch(input))
-            // {
-            //     ErrorMessage = "Name should not be less than or greater than 12 characters";
-            //     return false;
-            // }
-            // else if (!hasNumber.IsMatch(input))
-            // {
-            //     ErrorMessage = "Name should contain At least one numeric value";
-            //     return false;
-            // }
-
         }
 
-        public bool ValidateMinPrice(int? _minPrice, out string ErrorMessage)
+        public bool ValidateMinPrice(decimal? _minPrice, out string ErrorMessage)
         {
             ErrorMessage = string.Empty;
-            
-                if (_minPrice < 0)
-                {
-                    ErrorMessage = "Min Price should not be negative number";
-                    return false;
-                }
+
+            if (_minPrice < 0)
+            {
+                ErrorMessage = "Min Price should not be negative number";
+                return false;
+            }
 
             return true;
         }
 
-        public bool ValidateMaxPrice(int? _minPrice, int? _maxPrice, out string ErrorMessage)
+        public bool ValidateMaxPrice(decimal? _minPrice, decimal? _maxPrice, out string ErrorMessage)
         {
             ErrorMessage = string.Empty;
 
-                if (_maxPrice <= 0)
-                {
-                    ErrorMessage = "Max Price should not be negative number";
-                    return false;
-                }
-                else if (_maxPrice < _minPrice)
-                {
-                    ErrorMessage = "Max Price should not be less than Min Price";
-                    return false;
-                }
-
-
-
+            if (_maxPrice <= 0)
+            {
+                ErrorMessage = "Max Price should not be negative number";
+                return false;
+            }
+            else if (_maxPrice < _minPrice)
+            {
+                ErrorMessage = "Max Price should not be less than Min Price";
+                return false;
+            }
             return true;
         }
